pyproject.toml
[project]
name = "patent-chat"
version = "0.1.0"
description = "Streamlit app to ideate patents with Gemini and generate a draft specification"
authors = [{ name = "Your Name" }]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "streamlit>=1.37.1",
  "python-dotenv>=1.0.1",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "google-genai>=1.0.0",
  "python-docx>=1.1.2",
  "reportlab>=4.2.2",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.6",
  "pytest>=8.3.2",
  "pre-commit>=3.8.0",
]

[tool.uv]
# uv uses this table for resolution; empty is fine.

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.format]
quote-style = "preserve"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.per-file-ignores]
"app/main.py" = ["E402"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = false
